requirements.txt
# Function App dependencies only. Do not include azure-functions-worker.
azure-functions==1.25.0b3
requests==2.32.3
